requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1
pycodestyle==2.9.1
sqlparse==0.4.2
toml==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse
tzdata==2023.4
urllib3==2.1.0


